+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+## [0.2.1] - 2024-08-11
+
+Added:
+
+- Support `pnpm` package manager ([@skryukov])
+- New `--verbose` option for the installation generator ([@skryukov])
+
+Fixed:
+
+- Support installation alongside Webpacker ([@skryukov])
+
 ## [0.2.0] - 2024-08-10
 
 Added:
@@ -32,7 +43,8 @@ Added:
 [@iurev]: https://github.com/iurev
 [@skryukov]: https://github.com/skryukov
 
-[Unreleased]: https://github.com/skryukov/inertia_rails-contrib/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/skryukov/inertia_rails-contrib/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/skryukov/inertia_rails-contrib/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/skryukov/inertia_rails-contrib/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/skryukov/inertia_rails-contrib/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/skryukov/inertia_rails-contrib/commits/v0.1.0